Migrate to grunt-dojo2
