+++ .npmignore
@@ -1,7 +1,12 @@
+/_build
+/support
+/tests
+/typings
+.editorconfig
+.gitattributes
+.gitignore
 .npmignore
-.sublimets
 .travis.yml
-tests/
 Gruntfile.js
+tsconfig.json
 tslint.json
-*.ts
