+++ typings.json
@@ -1,0 +1,8 @@
+{
+	"name": "dojo-loader",
+	"dependencies": {},
+	"devDependencies": {},
+	"ambientDependencies": {
+		"node": "registry:dt/node#4.0.0+20160509154515"
+	}
+}
