+++ typings/browser.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="browser/ambient/node/index.d.ts" />
