+++ typings/main.d.ts
@@ -1,0 +1,8 @@
+/// <reference path="main/ambient/node/index.d.ts" />
+
+/* Simple Nashorn Typings */
+declare var arguments: any;
+declare var load: any;
+declare var Packages: any;
+
+declare var global: any;
