--- typings/nashorn/nashorn.d.ts
@@ -1,3 +1,0 @@
-declare var arguments: any;
-declare var load: any;
-declare var Packages: any;
