--- typings/tsd.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="nashorn/nashorn.d.ts" />
